+++ lib/src/flex_color_scheme.dart
@@ -6303,14 +6303,6 @@ class FlexColorScheme with Diagnosticable {
                       : math.min(subTheme.defaultRadius!, 10.0)),
               elevation: popupMenuElevation,
               color: popupMenuColor,
-              //
-              // useMaterial3
-              // TODO(rydmike): Remove tint elev temp fix when M3 supported.
-              //     ? ElevationOverlay.applySurfaceTint(
-              //         popupMenuColor ?? colorScheme.surface,
-              //         colorScheme.surfaceTint,
-              //         popupMenuElevation)
-              //     : popupMenuColor,
             )
           : null,
       dialogTheme: useSubThemes
@@ -6346,18 +6338,7 @@ class FlexColorScheme with Diagnosticable {
       bottomSheetTheme: useSubThemes
           ? FlexSubThemes.bottomSheetTheme(
               backgroundColor: bottomSheetColor,
-              //
-              // useMaterial3
-              // TODO(rydmike): Remove tint elev temp fix when M3 supported.
-              //     ? ElevationOverlay.applySurfaceTint(bottomSheetColor,
-              //         colorScheme.surfaceTint, bottomSheetElevation)
-              //     : bottomSheetColor,
               modalBackgroundColor: bottomSheetModalColor,
-              // useMaterial3
-              // TODO(rydmike): Remove tint elev temp fix when M3 supported.
-              //     ? ElevationOverlay.applySurfaceTint(bottomSheetModalColor,
-              //          colorScheme.surfaceTint, bottomSheetModalElevation)
-              //       : bottomSheetModalColor,
               elevation: bottomSheetElevation,
               modalElevation: bottomSheetModalElevation,
               radius: subTheme.bottomSheetRadius ?? subTheme.defaultRadius,